Finished Test Application For Phone

+++ Hololens/Hololens 2 - Phone Tracker 2020/Assets/ArUCO Markers/Scripts/UI/ArucoBoardLayoutItem.cs
@@ -22,7 +22,7 @@ public class ArucoBoardLayoutItem : SerializablePacket
     }
     public override void FromBinaryReader(EndianBinaryReader reader)
     {
-        size = reader.ReadInt32();
+        size = reader.ReadSingle();
         topLeftCorner = reader.ReadVector3();
     }
 }
